python-demo-app/pyproject.toml
[tool.poetry]
name = "python-demo-app"
version = "0.1.0"
description = ""
authors = ["Adam Nolte <ajnolte12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
autoblocksai = {git = "https://github.com/autoblocksai/python-sdk.git", branch = "adam/epd-2036-update-python-sdk"}
openai = "^1.76.0"
dotenv = "^0.9.9"
opentelemetry-instrumentation-openai = "^0.40.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.0"

[tool.poetry.scripts]
# This will execute the run() function in my_project/run.py
run_tests = "tests.test:run_tests"
create_dataset = "scripts.create_dataset:create"
create_prompts = "scripts.create_prompts:create"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
lint.select = ["E", "F", "I001"]
line-length = 120

[tool.ruff.lint.isort]
force-single-line = true
known-first-party = ["python_demo_app"]
